+++ src/TSV/TsvOfferDTO.java
@@ -1,8 +1,9 @@
 package TSV;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.Locale;
 
 public record TsvOfferDTO(Locale localization, String countryName, Date departureDate, Date arrivalDate, String place,
-                          Double price, String currencyCode) {
+                          BigDecimal price, String currencyCode) {
 }
